+++ sketch-compose-core/src/commonTest/kotlin/com/github/panpf/sketch/compose/core/common/test/AsyncImageTest.kt
@@ -256,7 +256,10 @@ class AsyncImageTest {
             }
             assertTrue(actual = stateHolder?.result is ImageResult.Success)
             assertEquals(
-                expected = IntSize(0, windowContainerSizeHolder!!.height),
+                expected = IntSize(
+                    windowContainerSizeHolder!!.width,
+                    windowContainerSizeHolder!!.height
+                ),
                 actual = (stateHolder?.result as ImageResult.Success).resize.size.toIntSize()
             )
         }
@@ -289,7 +292,10 @@ class AsyncImageTest {
             }
             assertTrue(actual = stateHolder?.result is ImageResult.Success)
             assertEquals(
-                expected = IntSize(windowContainerSizeHolder!!.width, 0),
+                expected = IntSize(
+                    windowContainerSizeHolder!!.width,
+                    windowContainerSizeHolder!!.height
+                ),
                 actual = (stateHolder?.result as ImageResult.Success).resize.size.toIntSize()
             )
         }
@@ -297,7 +303,9 @@ class AsyncImageTest {
         // width height is not bounded
         runComposeUiTest {
             var stateHolder: AsyncImageState? = null
+            var windowContainerSizeHolder: IntSize? = null
             setContent {
+                windowContainerSizeHolder = windowContainerSize()
                 LifecycleContainer {
                     Box(
                         modifier = Modifier.fillMaxSize().horizontalScroll(rememberScrollState())
@@ -321,7 +329,10 @@ class AsyncImageTest {
             }
             assertTrue(actual = stateHolder?.result is ImageResult.Success)
             assertEquals(
-                expected = IntSize(0, 0),
+                expected = IntSize(
+                    windowContainerSizeHolder!!.width,
+                    windowContainerSizeHolder!!.height
+                ),
                 actual = (stateHolder?.result as ImageResult.Success).resize.size.toIntSize()
             )
         }
@@ -370,7 +381,9 @@ class AsyncImageTest {
         // width is not bounded
         runComposeUiTest {
             var stateHolder: AsyncImageState? = null
+            var windowContainerSizeHolder: IntSize? = null
             setContent {
+                windowContainerSizeHolder = windowContainerSize()
                 LifecycleContainer {
                     Box(
                         modifier = Modifier.fillMaxSize().horizontalScroll(rememberScrollState())
@@ -393,7 +406,10 @@ class AsyncImageTest {
             }
             assertTrue(actual = stateHolder?.result is ImageResult.Success)
             assertEquals(
-                expected = IntSize(0, 0),
+                expected = IntSize(
+                    windowContainerSizeHolder!!.width,
+                    windowContainerSizeHolder!!.height
+                ),
                 actual = (stateHolder?.result as ImageResult.Success).resize.size.toIntSize()
             )
         }
@@ -401,7 +417,9 @@ class AsyncImageTest {
         // height is not bounded
         runComposeUiTest {
             var stateHolder: AsyncImageState? = null
+            var windowContainerSizeHolder: IntSize? = null
             setContent {
+                windowContainerSizeHolder = windowContainerSize()
                 LifecycleContainer {
                     Box(
                         modifier = Modifier.fillMaxSize().verticalScroll(rememberScrollState())
@@ -424,7 +442,10 @@ class AsyncImageTest {
             }
             assertTrue(actual = stateHolder?.result is ImageResult.Success)
             assertEquals(
-                expected = IntSize(0, 0),
+                expected = IntSize(
+                    windowContainerSizeHolder!!.width,
+                    windowContainerSizeHolder!!.height
+                ),
                 actual = (stateHolder?.result as ImageResult.Success).resize.size.toIntSize()
             )
         }
@@ -432,7 +453,9 @@ class AsyncImageTest {
         // width height is not bounded
         runComposeUiTest {
             var stateHolder: AsyncImageState? = null
+            var windowContainerSizeHolder: IntSize? = null
             setContent {
+                windowContainerSizeHolder = windowContainerSize()
                 LifecycleContainer {
                     Box(
                         modifier = Modifier.fillMaxSize().horizontalScroll(rememberScrollState())
@@ -456,7 +479,10 @@ class AsyncImageTest {
             }
             assertTrue(actual = stateHolder?.result is ImageResult.Success)
             assertEquals(
-                expected = IntSize(0, 0),
+                expected = IntSize(
+                    windowContainerSizeHolder!!.width,
+                    windowContainerSizeHolder!!.height
+                ),
                 actual = (stateHolder?.result as ImageResult.Success).resize.size.toIntSize()
             )
         }